removed std for root feature
