+++ Cargo.toml
@@ -44,10 +44,11 @@ nix = { version = "0.29", features = ["user"] }
 
 [features]
 all = ["macros", "network", "root", "std", "traits", "sized_string"]
+com = []
 default = ["all"]
 macros = ["quork-proc"]
-network = ["windows"]
-root = ["std"]
+network = ["windows", "com"]
+root = []
 sized_string = []
 std = []
 traits = []
